requirements.txt

Flask==3.0.0
Gunicorn==21.2.0
mysql-connector-python==8.2.0
pytz==2024.1
Flask-Cors==4.0.0
bcrypt==4.1.2
